ΩMEGA AGI SUPREMACY — Initial commit\n\nComplete 11-layer autonomous AGI system with:\n- Layer 0-10: core, agent, evolution, transcendece, superpowers\n- Multi-protocol adapters (OpenClaw, Hermes, OpenHuman, Feishu)\n- Self-evolving APEX architecture with φAPEX formula\n- Life harness with heartbeat, recovery, persistence\n- Research engine with web search + knowledge extraction\n- Docker + systemd + multi-platform deployment\n- Full CI/CD pipeline with cross-compilation\n\nCo-Authored-By: AtomCode (deepseek-v4-flash) <noreply@atomgit.com>

+++ omega-agi/superpowers/src/optimizer.rs
@@ -64,7 +64,7 @@ impl Optimizer {
         params.concurrency = (4 + (*count as u32 % 3)).max(2);
         params.batch_size = (16 + (*count as u32 % 8)).max(4);
         params.cache_ttl_secs = 300 + (*count as u64 % 60);
-        params.timeout_ms = (30000 - (*count as u32 % 5000)).max(5000);
+        params.timeout_ms = (30000 - (*count as u32 % 5000)).max(5000) as u64;
 
         let mut level = self.level.write().await;
         *level = (0.75 + improvement).min(0.99);
